#99: FindSearch View 파일 분리 및 코드 리팩토링
